Split iOS live smoke support package
