--- apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/LiveSmokeInteractions.swift
@@ -1,486 +1,0 @@
-import XCTest
-import Foundation
-
-private struct LiveSmokeTabBarItemCandidate {
-    let element: XCUIElement
-    let identifier: String
-    let note: String
-}
-
-extension LiveSmokeTestCase {
-    @MainActor
-    func tapTabBarItem(selectedTab: LiveSmokeSelectedTab, timeout: TimeInterval) throws {
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_tab.\(selectedTab.rawValue)") {
-            let tabBarItemLookup = selectedTab.tabBarItemLookup(localization: self.currentLaunchLocalization)
-            let tabBarItemCandidates = self.tabBarItemCandidates(lookup: tabBarItemLookup)
-            let deadline = Date().addingTimeInterval(timeout)
-
-            while Date() < deadline {
-                for candidate in tabBarItemCandidates {
-                    if candidate.element.exists && candidate.element.isHittable {
-                        try self.tapExistingElement(
-                            candidate.element,
-                            identifier: candidate.identifier,
-                            action: "tap_tab",
-                            note: candidate.note
-                        )
-                        return
-                    }
-                }
-
-                RunLoop.current.run(until: Date(timeIntervalSinceNow: liveSmokeFocusPollIntervalSeconds))
-            }
-
-            throw LiveSmokeFailure.missingElement(
-                identifier: self.tabBarLookupDescription(selectedTab: selectedTab, lookup: tabBarItemLookup),
-                timeoutSeconds: timeout,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-    }
-
-    @MainActor
-    func tapTabBarItem(identifier: String, timeout: TimeInterval) throws {
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_tab.\(identifier)") {
-            let tabBarItem = self.app.tabBars.descendants(matching: .any).matching(identifier: identifier).firstMatch
-            let deadline = Date().addingTimeInterval(timeout)
-
-            while Date() < deadline {
-                if tabBarItem.exists && tabBarItem.isHittable {
-                    try self.tapExistingElement(
-                        tabBarItem,
-                        identifier: identifier,
-                        action: "tap_tab",
-                        note: "tab bar item tapped"
-                    )
-                    return
-                }
-
-                RunLoop.current.run(until: Date(timeIntervalSinceNow: 0.2))
-            }
-
-            throw LiveSmokeFailure.missingElement(
-                identifier: identifier,
-                timeoutSeconds: timeout,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-    }
-
-    @MainActor
-    func tapTabBarItem(named name: String, timeout: TimeInterval) throws {
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_tab.\(name)") {
-            let tabBarButton = self.app.tabBars.buttons[name].firstMatch
-            let deadline = Date().addingTimeInterval(timeout)
-
-            while Date() < deadline {
-                if tabBarButton.exists && tabBarButton.isHittable {
-                    try self.tapExistingButton(
-                        tabBarButton,
-                        identifier: "tab.\(name)",
-                        action: "tap_tab",
-                        note: "tab bar button tapped"
-                    )
-                    return
-                }
-
-                RunLoop.current.run(until: Date(timeIntervalSinceNow: 0.2))
-            }
-
-            throw LiveSmokeFailure.missingElement(
-                identifier: "tab.\(name)",
-                timeoutSeconds: timeout,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-    }
-
-    @MainActor
-    private func tabBarItemCandidates(lookup: LiveSmokeTabBarItemLookup) -> [LiveSmokeTabBarItemCandidate] {
-        return [
-            LiveSmokeTabBarItemCandidate(
-                element: self.app.tabBars.buttons.matching(identifier: lookup.identifier).firstMatch,
-                identifier: lookup.identifier,
-                note: "tab bar button tapped via accessibility identifier"
-            ),
-            LiveSmokeTabBarItemCandidate(
-                element: self.app.tabBars.buttons[lookup.localizedTitle].firstMatch,
-                identifier: "tab.\(lookup.localizedTitle)",
-                note: "tab bar button tapped via localized label fallback"
-            ),
-            LiveSmokeTabBarItemCandidate(
-                element: self.app.tabBars.buttons.element(boundBy: lookup.stableIndex),
-                identifier: "tab.index.\(lookup.stableIndex)",
-                note: "tab bar button tapped via stable index fallback"
-            )
-        ]
-    }
-
-    @MainActor
-    private func tabBarLookupDescription(
-        selectedTab: LiveSmokeSelectedTab,
-        lookup: LiveSmokeTabBarItemLookup
-    ) -> String {
-        let visibleTabButtons = self.visibleTabBarButtonSummary()
-        return "tab.\(selectedTab.rawValue) tried identifier='\(lookup.identifier)' label='\(lookup.localizedTitle)' index=\(lookup.stableIndex) visibleButtons=[\(visibleTabButtons)]"
-    }
-
-    @MainActor
-    private func visibleTabBarButtonSummary() -> String {
-        let visibleButtons = self.elements(query: self.app.tabBars.buttons).enumerated().map { index, button in
-            let label = button.label.trimmingCharacters(in: .whitespacesAndNewlines)
-            let identifier = button.identifier.trimmingCharacters(in: .whitespacesAndNewlines)
-            let resolvedLabel = label.isEmpty ? "-" : label
-            let resolvedIdentifier = identifier.isEmpty ? "-" : identifier
-            return "\(resolvedLabel){index=\(index),id=\(resolvedIdentifier)}"
-        }
-
-        if visibleButtons.isEmpty {
-            return "<none>"
-        }
-
-        return visibleButtons.joined(separator: ", ")
-    }
-
-    @MainActor
-    func tapButton(identifier: String, timeout: TimeInterval) throws {
-        let button = self.app.buttons[identifier].firstMatch
-        try self.tapButton(button: button, identifier: identifier, timeout: timeout)
-    }
-
-    @MainActor
-    func tapButtonPreservingAlerts(identifier: String, timeout: TimeInterval) throws {
-        let button = self.app.buttons[identifier].firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_button_preserving_alerts.\(identifier)") {
-            if self.waitForOptionalElement(
-                button,
-                identifier: identifier,
-                timeout: timeout
-            ) == false {
-                throw LiveSmokeFailure.missingElement(
-                    identifier: identifier,
-                    timeoutSeconds: timeout,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            if button.isEnabled == false {
-                throw LiveSmokeFailure.disabledElement(
-                    identifier: identifier,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            self.logActionStart(action: "tap_button_preserving_alerts", identifier: identifier)
-            button.tap()
-            self.logActionEnd(
-                action: "tap_button_preserving_alerts",
-                identifier: identifier,
-                result: "success",
-                note: "button tapped without alert dismissal"
-            )
-        }
-    }
-
-    @MainActor
-    func tapButtonScrollingIntoView(identifier: String, timeout: TimeInterval) throws {
-        let button = self.app.buttons[identifier].firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_button_scrolling_into_view.\(identifier)") {
-            let deadline = Date().addingTimeInterval(timeout)
-
-            while Date() < deadline {
-                if button.exists && button.isHittable {
-                    try self.tapExistingButton(
-                        button,
-                        identifier: identifier,
-                        action: "tap_button_scrolling_into_view",
-                        note: "button tapped after scrolling into view"
-                    )
-                    return
-                }
-
-                self.scrollBestEffort()
-                RunLoop.current.run(until: Date(timeIntervalSinceNow: liveSmokeFocusPollIntervalSeconds))
-            }
-
-            throw LiveSmokeFailure.missingElement(
-                identifier: identifier,
-                timeoutSeconds: timeout,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-    }
-
-    @MainActor
-    func tapButtonScrollingIntoViewPreservingAlerts(identifier: String, timeout: TimeInterval) throws {
-        let button = self.app.buttons[identifier].firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_button_scrolling_into_view_preserving_alerts.\(identifier)") {
-            let deadline = Date().addingTimeInterval(timeout)
-
-            while Date() < deadline {
-                if button.exists && button.isHittable {
-                    try self.tapExistingButtonPreservingAlerts(
-                        button,
-                        identifier: identifier,
-                        action: "tap_button_scrolling_into_view_preserving_alerts",
-                        note: "button tapped after scrolling into view without alert dismissal"
-                    )
-                    return
-                }
-
-                self.scrollBestEffort()
-                RunLoop.current.run(until: Date(timeIntervalSinceNow: liveSmokeFocusPollIntervalSeconds))
-            }
-
-            throw LiveSmokeFailure.missingElement(
-                identifier: identifier,
-                timeoutSeconds: timeout,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-    }
-
-    @MainActor
-    func tapCell(identifier: String, timeout: TimeInterval) throws {
-        let cell = self.app.cells[identifier].firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_cell.\(identifier)") {
-            if self.waitForOptionalElement(
-                cell,
-                identifier: identifier,
-                timeout: timeout
-            ) == false {
-                throw LiveSmokeFailure.missingElement(
-                    identifier: identifier,
-                    timeoutSeconds: timeout,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            try self.tapExistingCell(cell, identifier: identifier)
-        }
-    }
-
-    @MainActor
-    func tapAlertButton(label: String, timeout: TimeInterval) throws {
-        let identifier = "alert.\(label)"
-        let button = self.app.alerts.buttons[label].firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_alert_button.\(label)") {
-            if self.waitForOptionalElement(
-                button,
-                identifier: identifier,
-                timeout: timeout
-            ) == false {
-                throw LiveSmokeFailure.missingElement(
-                    identifier: identifier,
-                    timeoutSeconds: timeout,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            try self.tapExistingButton(
-                button,
-                identifier: identifier,
-                action: "tap_alert_button",
-                note: "alert button tapped"
-            )
-        }
-    }
-
-    @MainActor
-    func tapAlertButtonPreservingAlerts(label: String, timeout: TimeInterval) throws {
-        let identifier = "alert.\(label)"
-        let button = self.app.alerts.buttons[label].firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_alert_button_preserving_alerts.\(label)") {
-            let deadline = Date().addingTimeInterval(timeout)
-            while Date() < deadline && button.exists == false {
-                RunLoop.current.run(until: Date(timeIntervalSinceNow: liveSmokeFocusPollIntervalSeconds))
-            }
-
-            if button.exists == false {
-                throw LiveSmokeFailure.missingElement(
-                    identifier: identifier,
-                    timeoutSeconds: timeout,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            if button.isEnabled == false {
-                throw LiveSmokeFailure.disabledElement(
-                    identifier: identifier,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            try self.tapExistingButtonPreservingAlerts(
-                button,
-                identifier: identifier,
-                action: "tap_alert_button_preserving_alerts",
-                note: "alert button tapped without dismissal"
-            )
-        }
-    }
-
-    @MainActor
-    func tapFirstNavigationBackButton() throws {
-        let backButton = self.app.navigationBars.buttons.firstMatch
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_back_button.navigation.backButton") {
-            if self.waitForOptionalElement(
-                backButton,
-                identifier: "navigation.backButton",
-                timeout: LiveSmokeConfiguration.shortUiTimeoutSeconds
-            ) == false {
-                throw LiveSmokeFailure.missingBackButton(
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            try self.tapExistingButton(
-                backButton,
-                identifier: "navigation.backButton",
-                action: "tap_back_button",
-                note: "back tapped"
-            )
-        }
-    }
-
-    @MainActor
-    func tapButton(
-        button: XCUIElement,
-        identifier: String,
-        timeout: TimeInterval
-    ) throws {
-        try self.runWithInlineRawScreenStateOnFailure(action: "tap_button.\(identifier)") {
-            if self.waitForOptionalElement(
-                button,
-                identifier: identifier,
-                timeout: timeout
-            ) == false {
-                throw LiveSmokeFailure.missingElement(
-                    identifier: identifier,
-                    timeoutSeconds: timeout,
-                    screen: self.currentScreenSummary(),
-                    step: self.currentStepTitle
-                )
-            }
-
-            try self.tapExistingButton(
-                button,
-                identifier: identifier,
-                action: "tap_button",
-                note: "button tapped"
-            )
-        }
-    }
-
-    @MainActor
-    private func tapExistingButton(
-        _ button: XCUIElement,
-        identifier: String,
-        action: String,
-        note: String
-    ) throws {
-        if button.isEnabled == false {
-            throw LiveSmokeFailure.disabledElement(
-                identifier: identifier,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-
-        self.logActionStart(action: action, identifier: identifier)
-        button.tap()
-        self.logActionEnd(action: action, identifier: identifier, result: "success", note: note)
-    }
-
-    @MainActor
-    private func tapExistingButtonPreservingAlerts(
-        _ button: XCUIElement,
-        identifier: String,
-        action: String,
-        note: String
-    ) throws {
-        if button.isEnabled == false {
-            throw LiveSmokeFailure.disabledElement(
-                identifier: identifier,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-
-        self.logActionStart(action: action, identifier: identifier)
-        button.tap()
-        self.logActionEnd(action: action, identifier: identifier, result: "success", note: note)
-    }
-
-    @MainActor
-    private func tapExistingElement(
-        _ element: XCUIElement,
-        identifier: String,
-        action: String,
-        note: String
-    ) throws {
-        if element.elementType == .button {
-            try self.tapExistingButton(
-                element,
-                identifier: identifier,
-                action: action,
-                note: note
-            )
-            return
-        }
-
-        self.logActionStart(action: action, identifier: identifier)
-        element.tap()
-        self.logActionEnd(action: action, identifier: identifier, result: "success", note: note)
-    }
-
-    @MainActor
-    private func scrollBestEffort() {
-        let collectionView = self.app.collectionViews.firstMatch
-        if collectionView.exists {
-            collectionView.swipeUp()
-            return
-        }
-
-        let scrollView = self.app.scrollViews.firstMatch
-        if scrollView.exists {
-            scrollView.swipeUp()
-            return
-        }
-
-        let table = self.app.tables.firstMatch
-        if table.exists {
-            table.swipeUp()
-            return
-        }
-
-        self.app.swipeUp()
-    }
-
-    @MainActor
-    private func tapExistingCell(_ cell: XCUIElement, identifier: String) throws {
-        if cell.isEnabled == false {
-            throw LiveSmokeFailure.disabledElement(
-                identifier: identifier,
-                screen: self.currentScreenSummary(),
-                step: self.currentStepTitle
-            )
-        }
-
-        self.logActionStart(action: "tap_cell", identifier: identifier)
-        cell.tap()
-        self.logActionEnd(action: "tap_cell", identifier: identifier, result: "success", note: "cell tapped")
-    }
-}
