+++ docs/xcode-cloud-data-access.md
@@ -91,8 +91,8 @@ Download URLs in `ciArtifacts.attributes.downloadUrl` are temporary pre-signed A
 This repository already emits structured smoke breadcrumbs from the iOS UI test suite:
 
 - grouped smoke entrypoints: [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeAiTests.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeAiTests.swift), [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeCardsTests.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeCardsTests.swift), [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeReviewTests.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeReviewTests.swift), and [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSettingsTests.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSettingsTests.swift)
-- shared smoke base: [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/LiveSmokeTestCase.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/LiveSmokeTestCase.swift)
-- breadcrumb logger and screen diagnostics: [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/LiveSmokeDiagnostics.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/LiveSmokeDiagnostics.swift)
+- shared smoke base: [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Configuration/LiveSmokeTestCase.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Configuration/LiveSmokeTestCase.swift)
+- breadcrumb logger and screen diagnostics: [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Diagnostics/LiveSmokeBreadcrumbs.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Diagnostics/LiveSmokeBreadcrumbs.swift), [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Diagnostics/LiveSmokeDiagnosticSnapshots.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Diagnostics/LiveSmokeDiagnosticSnapshots.swift), and [`apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Diagnostics/LiveSmokeElementState.swift`](../apps/ios/Flashcards/FlashcardsUITests/LiveSmokeSupport/Diagnostics/LiveSmokeElementState.swift)
 
 Those breadcrumbs appear in the cloud `xcodebuild` log as JSON lines under the `ios_ui_smoke` domain. That is the fastest way to recover exact per-step timestamps without opening the Xcode Cloud UI manually.
